+++ src/main/java/liza/main.java
@@ -1,0 +1,11 @@
+package liza;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class main {
+    public static void main(String[] args) {
+        SpringApplication.run(main.class,args);
+    }
+}
