--- src/main/java/main/java/Controler.java
@@ -1,25 +1,0 @@
-package main.java;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.servlet.ModelAndView;
-
-@Controller
-class Controler {
-
-        @RequestMapping(value = "/", method = RequestMethod.GET)
-        public ModelAndView allFilms() {
-            ModelAndView modelAndView = new ModelAndView();
-            modelAndView.setViewName("films");
-            return modelAndView;
-        }
-
-        @RequestMapping(value = "/edit", method = RequestMethod.GET)
-        public ModelAndView editPage() {
-            ModelAndView modelAndView = new ModelAndView();
-            modelAndView.setViewName("editPage");
-            return modelAndView;
-        }
-    }
-
