--- src/main/java/main/java/main.java
@@ -1,11 +1,0 @@
-package main.java;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class main {
-    public static void main(String[] args) {
-        SpringApplication.run(main.class,args);
-    }
-}
